+++ coverage/sqlitedb.py
@@ -76,13 +76,14 @@ class SqliteDb:
         # to keep things going.
         self.execute_void("pragma synchronous=off", fail_ok=True)
 
-    def close(self) -> None:
+    def close(self, force: bool = False) -> None:
         """If needed, close the connection."""
-        if self.con is not None and self.filename != ":memory:":
-            if self.debug.should("sql"):
-                self.debug.write(f"Closing {self.con!r} on {self.filename!r}")
-            self.con.close()
-            self.con = None
+        if self.con is not None:
+            if force or self.filename != ":memory:":
+                if self.debug.should("sql"):
+                    self.debug.write(f"Closing {self.con!r} on {self.filename!r}")
+                self.con.close()
+                self.con = None
 
     def __enter__(self) -> SqliteDb:
         if self.nest == 0: